generate and use ~/.config/JHTDB folder
the authentication token is now a file in this folder.
the interpolator class also stores all of its different files in this
folder as well (polynomial data, c files, shared libraries).
the main reason for the folder is the interpolator class, since it takes
a while for these things to be generated.